+++ src/checkmaite/core/object_detection/__init__.py
@@ -1,7 +1,7 @@
 from checkmaite.core._common.dataeval_bias_capability import DataevalBiasConfig, DataevalBiasOutputs
 from checkmaite.core._common.dataeval_cleaning_capability import DataevalCleaningConfig, DataevalCleaningOutputs
 from checkmaite.core._common.dataeval_shift_capability import DataevalShiftConfig, DataevalShiftOutputs
-from checkmaite.core._common.maite_evaluation_capability import MaiteEvaluationConfig, MaiteEvaluationOutputs
+from checkmaite.core._common.maite_evaluation_capability import MaiteEvaluationOutputs
 from checkmaite.core._common.nrtk_robustness_capability import NrtkRobustnessConfig, NrtkRobustnessOutputs
 from checkmaite.core._plugins import inject_plugin_exports
 from checkmaite.core.object_detection.dataeval_bias_capability import DataevalBias
@@ -12,7 +12,7 @@ from checkmaite.core.object_detection.dataeval_feasibility_capability import (
     DataevalFeasibilityOutputs,
 )
 from checkmaite.core.object_detection.dataeval_shift_capability import DataevalShift
-from checkmaite.core.object_detection.maite_evaluation_capability import MaiteEvaluation
+from checkmaite.core.object_detection.maite_evaluation_capability import MaiteEvaluation, MaiteEvaluationConfig
 from checkmaite.core.object_detection.nrtk_robustness_capability import NrtkRobustness
 from checkmaite.core.object_detection.xaitk_explainable_capability import (
     XaitkExplainable,